+++ modules/core/jvm/src/main/scala/com/azavea/stac4s/meta/HasInstant.scala
@@ -1,0 +1,20 @@
+package com.azavea.stac4s.meta
+
+import eu.timepit.refined.api.Validate
+
+import java.time.Instant
+
+final case class HasInstant()
+
+object HasInstant {
+
+  implicit def hasInstant: Validate.Plain[Option[Instant], HasInstant] =
+    Validate.fromPredicate(
+      {
+        case None => false
+        case _    => true
+      },
+      t => s"Value Is None: $t",
+      HasInstant()
+    )
+}